memory handling improved

+++ parse_vector.c
@@ -5,9 +5,9 @@
 
 #define SIZE INT_MAX // INT_MAX = +2147483647
 
-int *parse_vector(unsigned int n) {
-	int *vector = malloc(SIZE * sizeof(int));
-	int length = (int) floor(log10(n)); // log_base10(integer) -> integer lenght
+int16_t *parse_vector(int32_t n) {
+	int16_t *vector = malloc(SIZE * sizeof(int16_t));
+	int16_t length = (int) floor(log10(n)); // log_base10(integer) -> integer lenght
 	for (int i = 0; i <= length; i++) {
 		vector[length-i] = n % 10; // inserting elements from backward
 		n = n / 10;
@@ -17,7 +17,7 @@ int *parse_vector(unsigned int n) {
 
 int main() 
 {
-	int *vector = parse_vector(69420); // 69420 -> {6, 9, 4, 2, 0}
+	int16_t *vector = parse_vector(69420); // 69420 -> {6, 9, 4, 2, 0}
 	free(vector);
 	return 0;
 }
